fix(ci): maintainability shard — decompose three files back under the 1500-line limit
The post-PR protected-pause work pushed loop.py (1546), remote_repair.py (1574)
and test_pr_monitor_operator_hints.py (1557) over the first-party line limit,
failing tests/unit/test_core_decomposition_maintainability.py on the coverage
shard. Mechanically extract cohesive groups without behavior change:

- loop.py: move the protected/workflow-scope pause notification helpers into a
  new pause_notifications.py sibling.
- remote_repair.py: move the protected-scope push-block terminals into a new
  protected_scope_blocks.py sibling; rewire the mixin to bind them.
- test_pr_monitor_operator_hints.py: relocate the persisted-state round-trip /
  refresh tests into the existing test_pr_monitor_operator_hint_state.py.
